# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/contact_messages:
    get:
      tags:
        - ContactMessages
      summary: List ContactMessages
      description: Get a paginated list of ContactMessages with optional filtering and sorting
      operationId: listContactMessages
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactMessageList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ContactMessages
      summary: Create ContactMessage
      description: Create a new ContactMessage
      operationId: createContactMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactMessageRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/menu_blocks:
    get:
      tags:
        - MenuBlocks
      summary: List MenuBlocks
      description: Get a paginated list of MenuBlocks with optional filtering and sorting
      operationId: listMenuBlocks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuBlockList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MenuBlocks
      summary: Create MenuBlock
      description: Create a new MenuBlock
      operationId: createMenuBlock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuBlockRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuBlock'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    ContactMessage:
      type: object
      required:
        - id
        - website_id
        - email
        - message
        - notified
        - metadata
      properties:
        id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        name:
          nullable: true
          type: string
          maxLength: 120
        email:
          type: string
          maxLength: 320
        message:
          type: string
          maxLength: 5000
        notified:
          type: boolean
        metadata:
          type: object
...
    CreateContactMessageRequest:
      type: object
      required:
        - website_id
        - email
        - message
        - notified
        - metadata
      properties:
        website_id:
          type: string
          format: uuid
        name:
          nullable: true
          type: string
          maxLength: 120
        email:
          type: string
          maxLength: 320
        message:
          type: string
          maxLength: 5000
        notified:
          type: boolean
        metadata:
          type: object
...
    ContactMessageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContactMessage'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    MenuBlock:
      type: object
      required:
        - id
        - menu_id
        - kind
        - position
        - payload
      properties:
        id:
          type: string
          format: uuid
        menu_id:
          type: string
          format: uuid
        kind:
          type: string
        position:
          type: integer
          format: int32
        payload:
          type: object
...
    CreateMenuBlockRequest:
      type: object
      required:
        - menu_id
        - kind
        - position
        - payload
      properties:
        menu_id:
          type: string
          format: uuid
        kind:
          type: string
        position:
          type: integer
          format: int32
        payload:
          type: object
...
    MenuBlockList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuBlock'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []