# src/presentation/http/contact_message_handler.rs
//! ContactMessage REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{ContactMessageService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateContactMessageDto, UpdateContactMessageDto, PatchContactMessageDto, ContactMessageResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum ContactMessageError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for ContactMessageError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for ContactMessageError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "CONTACTMESSAGE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "CONTACTMESSAGE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "CONTACTMESSAGE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "CONTACTMESSAGE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for ContactMessage.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /contact_messages | List with pagination |
/// | POST | /contact_messages | Create new |
/// | GET | /contact_messages/:id | Get by ID |
/// | PUT | /contact_messages/:id | Full update |
/// | PATCH | /contact_messages/:id | Partial update |
/// | DELETE | /contact_messages/:id | Soft delete |
/// | POST | /contact_messages/bulk | Bulk create |
/// | POST | /contact_messages/upsert | Upsert |
/// | GET | /contact_messages/trash | List deleted |
/// | POST | /contact_messages/:id/restore | Restore |
/// | DELETE | /contact_messages/empty | Empty trash |
/// | GET | /contact_messages/:id/deleted | Get deleted by ID |
/// | DELETE | /contact_messages/trash/:id | Permanent delete from trash |
/// | GET | /contact_messages/count | Count active entities |
/// | GET | /contact_messages/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(ContactMessageService::with_repository(repository));
/// let router = create_contact_message_routes(service);
/// ```
pub fn create_contact_message_routes(service: Arc<ContactMessageService>) -> Router {
    BackboneCrudHandler::<ContactMessageService, ContactMessage, CreateContactMessageDto, UpdateContactMessageDto, ContactMessageResponseDto>::routes(
        service,
        "/contact_messages",
    )
}

/// Create Axum router with only the read (GET) endpoints for ContactMessage.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_contact_message_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_contact_message_read_routes(service: Arc<ContactMessageService>) -> Router {
    BackboneCrudHandler::<ContactMessageService, ContactMessage, CreateContactMessageDto, UpdateContactMessageDto, ContactMessageResponseDto>::read_routes(
        service,
        "/contact_messages",
    )
}

/// Create Axum router with only the write (mutation) endpoints for ContactMessage.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_contact_message_write_routes(service: Arc<ContactMessageService>) -> Router {
    BackboneCrudHandler::<ContactMessageService, ContactMessage, CreateContactMessageDto, UpdateContactMessageDto, ContactMessageResponseDto>::write_routes(
        service,
        "/contact_messages",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_contact_message_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<ContactMessageService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_contact_message_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}