# src/presentation/http/menu_block_handler.rs
//! MenuBlock REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{MenuBlockService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateMenuBlockDto, UpdateMenuBlockDto, PatchMenuBlockDto, MenuBlockResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum MenuBlockError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for MenuBlockError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for MenuBlockError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "MENUBLOCK_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "MENUBLOCK_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "MENUBLOCK_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "MENUBLOCK_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for MenuBlock.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /menu_blocks | List with pagination |
/// | POST | /menu_blocks | Create new |
/// | GET | /menu_blocks/:id | Get by ID |
/// | PUT | /menu_blocks/:id | Full update |
/// | PATCH | /menu_blocks/:id | Partial update |
/// | DELETE | /menu_blocks/:id | Soft delete |
/// | POST | /menu_blocks/bulk | Bulk create |
/// | POST | /menu_blocks/upsert | Upsert |
/// | GET | /menu_blocks/trash | List deleted |
/// | POST | /menu_blocks/:id/restore | Restore |
/// | DELETE | /menu_blocks/empty | Empty trash |
/// | GET | /menu_blocks/:id/deleted | Get deleted by ID |
/// | DELETE | /menu_blocks/trash/:id | Permanent delete from trash |
/// | GET | /menu_blocks/count | Count active entities |
/// | GET | /menu_blocks/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(MenuBlockService::with_repository(repository));
/// let router = create_menu_block_routes(service);
/// ```
pub fn create_menu_block_routes(service: Arc<MenuBlockService>) -> Router {
    BackboneCrudHandler::<MenuBlockService, MenuBlock, CreateMenuBlockDto, UpdateMenuBlockDto, MenuBlockResponseDto>::routes(
        service,
        "/menu_blocks",
    )
}

/// Create Axum router with only the read (GET) endpoints for MenuBlock.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_menu_block_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_menu_block_read_routes(service: Arc<MenuBlockService>) -> Router {
    BackboneCrudHandler::<MenuBlockService, MenuBlock, CreateMenuBlockDto, UpdateMenuBlockDto, MenuBlockResponseDto>::read_routes(
        service,
        "/menu_blocks",
    )
}

/// Create Axum router with only the write (mutation) endpoints for MenuBlock.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_menu_block_write_routes(service: Arc<MenuBlockService>) -> Router {
    BackboneCrudHandler::<MenuBlockService, MenuBlock, CreateMenuBlockDto, UpdateMenuBlockDto, MenuBlockResponseDto>::write_routes(
        service,
        "/menu_blocks",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_menu_block_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<MenuBlockService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_menu_block_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}